Update plugin version to 1.2.11 and compatibility for 2024.2 EAP
The changes in this commit include an update to the CD-Dark-Theme plugin version in the grade.properties file, and the compatibility for 2024.2 EAP/release in the CHANGELOG.md. The plugin now supports up to build 242.
Simplify build.gradle.kts by removing unnecessary blocks and imports
Refactored the build.gradle.kts file to improve code clarity by using direct assignments instead of the set method.